Move loader and saver (plantgl dependent AND unused) in legacy. Fix setup.py s. Move lpy controls into plantlab.
