+++ plantlab/src/openalea/plantlab/picklable_curves.py
